fix(ui): explicitly convert baseRadiusDp to Float in RoundedKeyboardFrameView

+++ app/src/main/java/helium314/keyboard/latin/RoundedKeyboardFrameView.kt
@@ -143,7 +143,7 @@ class RoundedKeyboardFrameView @JvmOverloads constructor(
         val baseRadiusDp = Settings.readKeyboardCornerRadius(context.prefs()).coerceIn(
             Settings.KEYBOARD_CORNER_RADIUS_MIN_DP,
             Settings.KEYBOARD_CORNER_RADIUS_MAX_DP
-        )
+        ).toFloat()
         val radiusDp = if (isFloatingMode) maxOf(20f, baseRadiusDp) else baseRadiusDp
         return radiusDp * resources.displayMetrics.density
     }
